Remove redundant 'standard' test-classes
