+++ src/main/resources/api.yaml
@@ -129,11 +129,11 @@ components:
       type: object
       description: Example of a Record with default fields
       properties:
-        nullableStringField:
+        field1:
           type: string
           description: a nullable String field
           nullable: true
-        defaultStringField:
+        field2:
           type: string
           description: a String field with a default value
           default: someDefaultValue
