+++ src/test/java/com/chrimle/example/standard/StandardTests.java
@@ -4,6 +4,8 @@ import com.chrimle.example.utils.GeneratedEnumTestUtils;
 import com.chrimle.example.utils.GeneratedRecordTestUtils;
 import java.math.BigDecimal;
 import java.util.List;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 /**
@@ -121,5 +123,38 @@ public class StandardTests {
         List.class,
         List.class
     );
+    GeneratedRecordTestUtils.assertRecord(
+        exampleRecordWithDefaultFields,
+        false,
+        HAS_ADDITIONAL_MODEL_TYPE_ANNOTATIONS,
+        IS_MODEL_SERIALIZABLE,
+        String.class,
+        String.class
+    );
   }
+
+  @Test
+  @DisplayName("Testing nullable field is null after instantiation")
+  public void testNullableFieldIsNullAfterInstantiation() {
+    ExampleRecordWithDefaultFields record = new ExampleRecordWithDefaultFields(
+        null, "ignore");
+    Assertions.assertNull(record.field1());
+  }
+
+  @Test
+  @DisplayName("Testing field with default value is set to default if null is provided")
+  public void testFieldWithDefaultValueIsSetToDefaultIfProvidedValueIsNull() {
+    ExampleRecordWithDefaultFields record = new ExampleRecordWithDefaultFields(
+        "ignore", null);
+    Assertions.assertEquals("someDefaultValue", record.field2());
+  }
+
+  @Test
+  @DisplayName("Testing field with default value is set to provided value if not null")
+  public void testFieldWithDefaultValueIsSetToProvidedValueIfProvidedValueIsNotNull() {
+    ExampleRecordWithDefaultFields record = new ExampleRecordWithDefaultFields(
+        "ignore", "someValue");
+    Assertions.assertEquals("someValue", record.field2());
+  }
+
 }
