--- src/test/java/com/chrimle/example/standard/TestExampleRecordWithArrayFields.java
@@ -1,31 +1,0 @@
-package com.chrimle.example.standard;
-
-import com.chrimle.example.utils.AssertionUtils;
-import java.util.List;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-@DisplayName("Testing Record with Array fields")
-public class TestExampleRecordWithArrayFields {
-
-  private final Class<?> classUnderTest = ExampleRecordWithArrayFields.class;
-
-  @Test
-  @DisplayName("Testing class generated is a Record")
-  public void testClassGeneratedIsRecord() {
-    AssertionUtils.assertIsRecord(classUnderTest);
-  }
-
-  @Test
-  @DisplayName("Testing generated Record is not annotated as @Deprecated")
-  public void testRecordIsNotAnnotatedAsDeprecated() {
-    AssertionUtils.assertClassIsAnnotatedAsDeprecated(classUnderTest, false);
-  }
-
-  @Test
-  @DisplayName("Testing generated Record has expected number of fields of expected types")
-  public void testRecordHasNumberOfFieldsOfType() {
-    AssertionUtils.assertRecordHasNumberOfFieldsOfType(
-        classUnderTest, 3, List.class);
-  }
-}
