--- src/test/java/com/chrimle/example/standard/TestExampleRecordWithExampleEnumFields.java
@@ -1,30 +1,0 @@
-package com.chrimle.example.standard;
-
-import com.chrimle.example.utils.AssertionUtils;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-@DisplayName("Testing Record with ExampleEnum fields")
-public class TestExampleRecordWithExampleEnumFields {
-
-  private final Class<?> classUnderTest = ExampleRecordWithExampleEnumFields.class;
-
-  @Test
-  @DisplayName("Testing class generated is a Record")
-  public void testClassGeneratedIsRecord() {
-    AssertionUtils.assertIsRecord(classUnderTest);
-  }
-
-  @Test
-  @DisplayName("Testing generated Record is not annotated as @Deprecated")
-  public void testRecordIsNotAnnotatedAsDeprecated() {
-    AssertionUtils.assertClassIsAnnotatedAsDeprecated(classUnderTest, false);
-  }
-
-  @Test
-  @DisplayName("Testing generated Record has expected number of fields of expected types")
-  public void testRecordHasNumberOfFieldsOfType() {
-    AssertionUtils.assertRecordHasNumberOfFieldsOfType(
-        classUnderTest, 3, ExampleEnum.class);
-  }
-}
